+++ app/src/main/java/com/techcourse/AppWebApplicationInitializer.java
@@ -5,8 +5,8 @@ import org.slf4j.LoggerFactory;
 
 import jakarta.servlet.ServletContext;
 import nextstep.mvc.DispatcherServlet;
-import nextstep.mvc.controller.tobe.AnnotationHandlerAdapter;
-import nextstep.mvc.controller.tobe.AnnotationHandlerMapping;
+import nextstep.mvc.controller.AnnotationHandlerAdapter;
+import nextstep.mvc.controller.AnnotationHandlerMapping;
 import nextstep.web.WebApplicationInitializer;
 
 public class AppWebApplicationInitializer implements WebApplicationInitializer {
